- Finished Map definition function
- Created header for function prototypes to clean up main file

- Fixed logic error in read_input

-Commented through the read_input file

+++ robot-algorithm-simulator/global.h
@@ -2,6 +2,6 @@
 #ifndef GLOBAL_H
 #define GLOBAL_H
 
-extern int board_size;
+extern double board_size;
 
 #endif
